[2023][1.2] Broken
This code doesn't work for the actual data.
For all the test data, this is functional.
However the test data doesn't include certain cases which were happily provided by reddit.
This are "eighthree" = 83 and "sevenine" = 79.
Meaning that the solution needs to check the letters rather than convert to the integer and then check.